refactor service request to meal request, build prototype 1 jar
