+++ src/main/java/edu/wpi/romanticraijuu/fields/ServiceRequestData.java
@@ -1,4 +1,4 @@
-package edu.wpi.romanticraijuu;
+package edu.wpi.romanticraijuu.fields;
 
 import java.util.*;
